# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'todo list'
  description: 'todo list'
  version: 1.0.0
servers:
  -
    url: 'http://127.0.0.1:8000'
paths:
...
  '/api/task/{id}':
    get:
      summary: 'Display the specified resource.'
      operationId: displayTheSpecifiedResource
      description: ''
      parameters:
        -
          in: header
          name: Authorization
          description: ''
          example: 'Bearer {YOUR_AUTH_KEY}'
          schema:
            type: string
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  data:
                    id: 1
                    task_descr: 'buy coffee'
                    member_id: '9'
                    member_name: 公孫思宗
                    member_group: APP
                    creator_id: '9'
                    created_at: '2023-08-21 12:04:34'
                    updated_at: '2023-08-21 14:15:20'
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
                      task_descr:
                        type: string
                        example: 'buy coffee'
                      member_id:
                        type: string
                        example: '9'
                      member_name:
                        type: string
                        example: 公孫思宗
                      member_group:
                        type: string
                        example: APP
                      creator_id:
                        type: string
                        example: '9'
                      created_at:
                        type: string
                        example: '2023-08-21 12:04:34'
                      updated_at:
                        type: string
                        example: '2023-08-21 14:15:20'
      tags:
        - 'Task 任務資料'
    put:
      summary: ''
      operationId: putApiTaskId
      description: ''
      parameters:
        -
          in: header
          name: Authorization
          description: ''
          example: 'Bearer {YOUR_AUTH_KEY}'
          schema:
            type: string
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  data:
                    id: 1
                    task_descr: reprehenderit
                    member_id: '9'
                    member_name: 公孫思宗
                    member_group: APP
                    creator_id: '9'
                    created_at: '2023-08-21 12:04:34'
                    updated_at: '2023-08-22 10:26:59'
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
                      task_descr:
                        type: string
                        example: reprehenderit
                      member_id:
                        type: string
                        example: '9'
                      member_name:
                        type: string
                        example: 公孫思宗
                      member_group:
                        type: string
                        example: APP
                      creator_id:
                        type: string
                        example: '9'
                      created_at:
                        type: string
                        example: '2023-08-21 12:04:34'
                      updated_at:
                        type: string
                        example: '2023-08-22 10:26:59'
      tags:
        - 'Task 任務資料'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                task_descr:
                  type: string
                  description: 'The context of the task description.'
                  example: reprehenderit
                member_id:
                  type: integer
                  description: 'The id of the member who is asign to the task.'
                  example: 9
                creator_id:
                  type: integer
                  description: 'The id of the user who creat the task.'
                  example: 9
              required:
                - task_descr
                - member_id
                - creator_id
    delete:
      summary: 'Remove the specified resource from storage.'
      operationId: removeTheSpecifiedResourceFromStorage
      description: ''
      parameters:
        -
          in: header
          name: Authorization
          description: ''
          example: 'Bearer {YOUR_AUTH_KEY}'
          schema:
            type: string
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  Status: success
                  Message: 刪除成功
                properties:
                  Status:
                    type: string
                    example: success
                  Message:
                    type: string
                    example: 刪除成功
      tags:
        - 'Task 任務資料'
    parameters:
      -
        in: path
        name: id
        description: 'The ID of the task.'
        example: 1
        required: true
        schema:
          type: integer
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: ''
security:
  -
    default: []